🧹 `Marketplace`: `DeliveryArea` Strip TurboStreams
Now that we have a better idea how to use TurboFrames, the Streams seem
pretty unnecessary.

This also makes it so following links within the
`Marketplace::ManagementComponent` update the browser URL.

+++ app/furniture/marketplace/delivery_area_component.rb
@@ -33,9 +33,7 @@ class Marketplace
 
       ButtonComponent.new(label: "#{t("icons.discard")} #{t("discard.link_to")}",
         title: t("marketplace.delivery_areas.discard.link_to", name: delivery_area.label),
-        href: delivery_area.location, turbo_stream: true,
-        method: :delete,
-        scheme: :secondary)
+        href: delivery_area.location, method: :delete, scheme: :secondary)
     end
 
     def discard_button?
@@ -47,9 +45,7 @@ class Marketplace
 
       ButtonComponent.new(label: "#{t("icons.destroy")} #{t("destroy.link_to")}",
         title: t("marketplace.delivery_areas.destroy.link_to", name: delivery_area.label),
-        href: delivery_area.location, turbo_stream: true,
-        method: :delete,
-        confirm: t("destroy.confirm"),
+        href: delivery_area.location, method: :delete, confirm: t("destroy.confirm"),
         scheme: :secondary)
     end
